Implemented basic framework for a Ray Tracer
Framework includes a few Geometry objects (color, ray, and vec3), Scene
objects which will be elaborated on in the future (camera, world,
intersection), and Shapes (shape, sphere [unimplemented...], and
point_light). The Scene and Shape objects will be extended in much more
detail in the future, as soon as the actual tracing code is implemented
for the camera.

What I learned about Rust:
- Box<T> versus &T: There are two main ways to transfer a pointer to a
specific block of memory in Rust for use in functions, objects, etc.
These are Box<T> and &T. When you pass a parameter as Box<T>, you
essentially relinquish ownership of the memory, and it becomes the
responsibility of the new owner of the Box to free the memory when it's
out of usage. Rust will do this for you, but it is important to note
that once the new method or object owns the memory, as soon as that
method returns or the object is deleted, the memory is gone. On the
other hand, when you pass memory as &T, you pass it as a reference which
"borrows" the memory's ownership for a temporary length of time. When
the method which has this borrowed memory returns, the memory is not
deleted from the heap like a Box<T> would be; instead, ownership is
passed back down to the original owner of the memory. I will probably be
getting into some fun with Lifetimes as soon as I begin dealing with
actual iterations over scene objects, but for now the borrow checker
isn't giving me too much grief.
- Option<T> is a very interesting paradigm. Instead of returning a
"null" object or something of the sort, Rust is akin to using the
Option<T> type. This is basically an enumerable type which may or may
not have a value, which is accessed using Rust's "match" syntax. This is
useful for the Shape trait's intersect methods, which return
Option<Intersection>. This is because a Shape may not intersect a Ray,
and we cannot return null or another dummy value in that case.
- I've also learned a lot about the cargo/module structure in Rust. The
way I have the modules organized gives a bit of an odd structure for
"use" includes, for example "use shape::shape::Shape"... so I might have
to do file restructuring when I finally get around to studying the
best-practice file structuring in Rust.
- Finally, operator overloading is much more difficult to do in Rust.
It's by no means impossible, but it is a lot more boiler plate than
simply implementing an operator+() function in C++...